L2-365: Cleaner code
