+++ frontend/svelte/src/lib/components/neuron-detail/NeuronMetaInfoCard.svelte
@@ -12,6 +12,7 @@
     formatVotingPower,
     hasJoinedCommunityFund,
     isCurrentUserController,
+    isNeuronControllable,
   } from "../../utils/neuron.utils";
   import NeuronCard from "../neurons/NeuronCard.svelte";
   import Tooltip from "../ui/Tooltip.svelte";
@@ -21,7 +22,6 @@
   import SplitNeuronButton from "./actions/SplitNeuronButton.svelte";
   import DissolveActionButton from "./actions/DissolveActionButton.svelte";
   import DisburseButton from "./actions/DisburseButton.svelte";
-  import { isNeuronControllable } from "../../services/neurons.services";
   import { accountsStore } from "../../stores/accounts.store";
 
   export let neuron: NeuronInfo;
@@ -31,9 +31,10 @@
   let userControlled: boolean;
   $: userControlled = isCurrentUserController(neuron);
   let isControllable: boolean;
-  // Hack to call isNeuronControllable every time accounts store changes
-  // TODO: Move logic to `accountsStore` directly
-  $: $accountsStore, (isControllable = isNeuronControllable(neuron));
+  $: isControllable = isNeuronControllable({
+    neuron,
+    accounts: $accountsStore,
+  });
 </script>
 
 <NeuronCard {neuron}>
